requirements.txt
# CV Parsing & Text Processing
PyPDF2>=3.0.0
python-docx>=1.1.0
pdfplumber>=0.10.0

# NLP & Text Processing
spacy>=3.7.0
nltk>=3.8.0

# AI Model (LayoutLM için) - advanced_ai_extractor.py
transformers>=4.35.0
torch>=2.0.0
pillow>=10.0.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# API Framework
fastapi>=0.104.0
pydantic>=2.5.0
uvicorn>=0.24.0

# Utilities
python-multipart>=0.0.6
python-dotenv>=1.0.0
